requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
pywin32==306; sys_platform == 'win32'